SCF_Assignment/SCF_Session-4/SCF_Session-4_Q1/LargestMirrorSectionTest.java
import static org.junit.Assert.*;

import org.junit.Test;


public class LargestMirrorSectionTest {

	@Test
	public void testPass() {
		int elementsArray[]={1,2,3,8,9,3,2,1};
		LargestMirrorSection obj=new LargestMirrorSection();
		int result=obj.findLargestMirrorSection(elementsArray);
		
		assertEquals(3,result);
	}
	
	@Test
	public void testFail() {
		int elementsArray[]={1,2,3,8,9,3,2,1};
		LargestMirrorSection obj=new LargestMirrorSection();
		int result=obj.findLargestMirrorSection(elementsArray);
		
		assertEquals(4,result);
	}
	
	@Test
	public void emptyArrayTest() {
		int elementsArray[]={};
		LargestMirrorSection obj=new LargestMirrorSection();
		int result=obj.findLargestMirrorSection(elementsArray);
		
		assertEquals(4,result);
	}
	

	@Test
	public void testFail1() {
		int elementsArray[]={1,2,3,8,9,3,2,1};
		LargestMirrorSection obj=new LargestMirrorSection();
		int result=obj.findLargestMirrorSection(elementsArray);
		
		assertNotEquals(4,result);
	}
}
